refactor: 로그인 파일 구조 수정 및 kakao auth 적용
