+++ src/containers/Auth/Login/Login.module.scss
@@ -3,8 +3,6 @@
   flex-direction: column;
   align-items: center;
   gap: 50px;
-  width: 100%;
-  height: 100vh;
   padding: 50px 0;
 
   @media screen and (max-width: $md-breakpoint) {
@@ -13,8 +11,8 @@
 }
 
 .image {
-  width: 100%;
-  height: 80%;
+  width: 500px;
+  height: 500px;
   border-radius: 10px;
   background-color: $gray-300;
 }
