+++ src/containers/index.ts
@@ -1,1 +1,2 @@
+export * from './Auth';
 export * from './Main';
